// src/interfaces/video.ts
// Generated by https://quicktype.io

export interface Video {
  error: string;
  limit: number;
  offset: number;
  number_of_page_results: number;
  number_of_total_results: number;
  status_code: number;
  results: VideoResult[];
  version: string;
}

export interface VideoResult {
  api_detail_url: string;
  associations: Association[];
  deck: string;
  embed_player: string;
  guid: string;
  id: number;
  length_seconds: number;
  name: string;
  premium: boolean;
  publish_date: Date;
  site_detail_url: string;
  image: Image;
  user: User;
  hosts: null;
  crew: null;
  video_type: VideoType | null;
  video_show: VideoShow;
  video_categories: Association[];
  saved_time: null;
  youtube_id: null | string;
  low_url: null | string;
  high_url: null | string;
  url: string;
}

export interface Association {
  api_detail_url: string;
  guid?: string;
  id: number;
  name: string;
  site_detail_url: string;
}

export interface Image {
  icon_url: string;
  medium_url: string;
  screen_url: string;
  screen_large_url: string;
  small_url: string;
  super_url: string;
  thumb_url: string;
  tiny_url: string;
  original_url: string;
  image_tags?: ImageTags;
}

export enum ImageTags {
  AllImages = "All Images",
}

export enum User {
  Janman = "janman",
  Turboshawn = "turboshawn",
}

export interface VideoShow {
  api_detail_url: string;
  id: number;
  title: string;
  position: number;
  site_detail_url: string;
  image: Image;
  logo: Image | null;
}

export enum VideoType {
  Events = "Events",
  Features = "Features",
  GiantBombcast = "Giant Bombcast",
  Premium = "Premium",
}